implementacao listagem de compradores vinda do banco de dados
