+++ .gitignore
@@ -8,6 +8,8 @@
 # testing
 /coverage
 
+.env
+
 # production
 /build
