+++ public/env.js
@@ -1,0 +1,3 @@
+window.env = {
+  "URL_BACKEND": "https://localhost:7171"
+};
