--- src/pages/Compradores/Comprador.module.css
@@ -1,6 +1,0 @@
-.main{
-    display: flex;
-    /* justify-content: space-between; */
-    width: 100%;
-    height: 100vh;
-  }
