+++ src/pages/Compradores/CompradorForm/CompradorForm.module.css
@@ -1,3 +1,10 @@
+.main{
+  display: flex;
+  /* justify-content: space-between; */
+  width: 100%;
+  height: 100vh;
+}
+
 .container{
     display: flex;
     flex-direction: column;
