--- src/pages/Compradores/CompradorLista/index.js
@@ -1,1 +1,0 @@
-export { default } from './CompradorLista'
